Code refactor to consider DDD in Authentication Management Controller.
